refactor: fold fetchSubscriptionPlans back into actions/payment.ts
Keeps all payment-related action creators in a single file, matching the
reducer/selector consolidation already done for the payment slice.
